+++ crate/audio_play/src/system.rs
@@ -1,0 +1,3 @@
+pub use self::sequence_audio_play_system::SequenceAudioPlaySystem;
+
+mod sequence_audio_play_system;