+++ crate/character_play/src/sequence_handler/common/input.rs
@@ -1,10 +1,8 @@
 pub(crate) use self::{
-    run_stop_check::RunStopCheck, stand_x_movement_check::StandXMovementCheck,
-    stand_z_movement_check::StandZMovementCheck, walk_x_movement_check::WalkXMovementCheck,
-    walk_z_movement_check::WalkZMovementCheck,
+    stand_x_movement_check::StandXMovementCheck, stand_z_movement_check::StandZMovementCheck,
+    walk_x_movement_check::WalkXMovementCheck, walk_z_movement_check::WalkZMovementCheck,
 };
 
-mod run_stop_check;
 mod stand_x_movement_check;
 mod stand_z_movement_check;
 mod walk_x_movement_check;